requirements.txt
asgiref==3.6.0
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==4.2
django-cors-headers==3.14.0
django-countries==7.5.1
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.1.1
pytz==2023.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2023.3
zope.interface==6.0
